+++ pkg/api-server/resource_endpoints_test.go
@@ -20,7 +20,7 @@ import (
 	sample_model "github.com/Kong/kuma/pkg/test/resources/apis/sample"
 )
 
-var _ = Describe("Resource WS", func() {
+var _ = Describe("Resource Endpoints", func() {
 	var apiServer *api_server.ApiServer
 	var resourceStore store.ResourceStore
 	var client resourceApiClient
@@ -116,6 +116,42 @@ var _ = Describe("Resource WS", func() {
 				"type": "SampleTrafficRoute",
 				"name": "tr-2",
 				"mesh": "default",
+				"path": "/sample-path"
+			}`
+			body, err := ioutil.ReadAll(response.Body)
+			Expect(err).ToNot(HaveOccurred())
+			Expect(body).To(Or(
+				MatchJSON(fmt.Sprintf(`{"items": [%s,%s]}`, json1, json2)),
+				MatchJSON(fmt.Sprintf(`{"items": [%s,%s]}`, json2, json1)),
+			))
+		})
+
+		It("should list resources from all meshes", func() {
+			// given
+			putSampleResourceIntoStore(resourceStore, "tr-1", "mesh-1")
+			putSampleResourceIntoStore(resourceStore, "tr-2", "mesh-2")
+
+			// when
+			client = resourceApiClient{
+				address: apiServer.Address(),
+				path:    "/sample-traffic-routes",
+			}
+			response := client.list()
+
+			// then
+			Expect(response.StatusCode).To(Equal(200))
+			json1 := `
+			{
+				"type": "SampleTrafficRoute",
+				"name": "tr-1",
+				"mesh": "mesh-1",
+				"path": "/sample-path"
+			}`
+			json2 := `
+			{
+				"type": "SampleTrafficRoute",
+				"name": "tr-2",
+				"mesh": "mesh-2",
 				"path": "/sample-path"
 			}`
 			body, err := ioutil.ReadAll(response.Body)